Make middleware use ApiRequest rather than Request
